chore: clean up file structure
