+++ src/client/components/personalization/productCard.module.scss
@@ -1,0 +1,103 @@
+/**
+* Product Card
+*/
+.productCard {
+  border-radius: 8px;
+  padding: rem(16px);
+  border: 1px solid v('gray-box-stroke');
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(245px, 1fr));
+  gap: rem(24px);
+
+  @include media('<=phone') {
+    grid-template-columns: 1fr;
+    padding: rem(12px);
+  }
+}
+
+.productCardImage {
+  max-width: 100%;
+  object-fit: cover;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.productCardBody {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.productCardName {
+  color: v('dark-black');
+  font-size: rem(24px);
+  font-weight: 500;
+  line-height: 130%;
+
+  @include media('<=phoneLandscape') {
+    font-size: rem(20px);
+  }
+}
+
+.productCardSize {
+  color: v('dark-black');
+  font-size: rem(16px);
+  font-weight: 300;
+  line-height: 145%; /* 23.2px */
+}
+
+.productCardBodyBottomRow {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  gap: rem(8px);
+}
+
+.productCardPrice {
+  color: v('orange-gradient');
+  font-size: rem(16px);
+  line-height: 160%;
+  letter-spacing: 0.16px;
+}
+
+.productCardButtons {
+  display: flex;
+  gap: rem(8px);
+  flex-wrap: wrap;
+}
+
+.addToFavorites {
+  padding: 8px !important;
+}
+
+.productCardItemCounter {
+  display: flex;
+  padding: 0px 6px;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  border-radius: 4px;
+  border: 1px solid v('orange-gradient');
+
+  color: v('orange-gradient');
+  font-size: 14px;
+  font-weight: 600;
+  line-height: 150%;
+  height: 37px;
+
+  @include media('<=phone') {
+    gap: 4px;
+  }
+
+  div {
+    cursor: pointer;
+    display: flex;
+    stroke: v('orange-gradient');
+    padding: 8px 10px;
+
+    @include media('<=phone') {
+      padding: 6px 8px;
+    }
+  }
+}
