+++ src/pages/personalization/index.tsx
@@ -1,8 +1,8 @@
-import { FunctionComponent, useEffect, useState } from 'react';
-import { useDebounce } from 'react-use';
+import { useEffect, useState } from 'react';
+import { useDebounce, useSessionStorage } from 'react-use';
 import { useSearchHistory } from '../../client/api/personalization/use-search-history';
 import { UseCaseWrapper } from '../../client/components/common/UseCaseWrapper/UseCaseWrapper';
-import { CircularProgress } from '@mui/material';
+import { CircularProgress, Dialog, DialogActions, DialogContent, DialogContentText, DialogTitle } from '@mui/material';
 import { useProducts } from '../../client/api/personalization/use-products';
 import { useVisitorData } from '../../client/use-visitor-data';
 import { usePersonalizationNotification } from '../../client/hooks/personalization/use-personalization-notification';
@@ -16,160 +16,17 @@ import { CustomPageProps } from '../_app';
 import styles from './personalization.module.scss';
 import Image from 'next/image';
 import Button from '../../client/components/common/Button/Button';
-import HeartIcon from './img/heart.svg';
-import SearchIcon from './img/search.svg';
-import CartIcon from './img/cart.svg';
-import { UserCartItem } from '../../server/personalization/database';
-import { ButtonPlusSvg } from '../../client/img/buttonPlusSvg';
-import { ButtonMinusSvg } from '../../client/img/buttonMinusSvg';
+import CartIcon from '../../client/img/cart.svg';
 import { Cart, CartProduct } from '../../client/components/common/Cart/Cart';
-
-type SearchProps = {
-  search: string;
-  setSearch: React.Dispatch<React.SetStateAction<string>>;
-};
-const Search: FunctionComponent<SearchProps> = ({ search, setSearch }) => {
-  return (
-    <div>
-      <div className={styles.searchBox}>
-        <Image src={SearchIcon} alt="Search" className={styles.searchIcon} />
-        <input
-          type="text"
-          className={styles.searchInput}
-          placeholder="Search for your favorite coffee"
-          onChange={(e) => setSearch(e.target.value)}
-          value={search}
-        />
-      </div>
-    </div>
-  );
-};
-
-const SEARCH_HISTORY_DISPLAY_LIMIT = 6;
-
-type SearchHistoryProps = {
-  searchHistory: string[];
-  setSearchHistory: (searchTerm: string) => void;
-};
-
-const SearchHistory: FunctionComponent<SearchHistoryProps> = ({ searchHistory, setSearchHistory }) => {
-  if (searchHistory.length === 0) {
-    return null;
-  }
-
-  return (
-    <div>
-      <div className={styles.searchHistory}>
-        Last searches:{' '}
-        {searchHistory.slice(0, SEARCH_HISTORY_DISPLAY_LIMIT).map((searchTerm, index) => (
-          <>
-            <span key={index} onClick={() => setSearchHistory(searchTerm)} className={styles.searchTerm}>
-              {searchTerm}
-            </span>
-            ,{' '}
-          </>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-type Product = {
-  price: number;
-  name: string;
-  image: string;
-  id: number;
-};
-
-const ItemCounter: FunctionComponent<{
-  count: number;
-  addItem: () => void;
-  removeItem: () => void;
-}> = ({ count, addItem, removeItem }) => {
-  return (
-    <div className={styles.productCardItemCounter}>
-      <div onClick={removeItem}>
-        <ButtonMinusSvg />
-      </div>
-      <span>{count}</span>
-      <div onClick={addItem}>
-        <ButtonPlusSvg />
-      </div>
-    </div>
-  );
-};
-
-const ProductCard: FunctionComponent<{ product: Product }> = ({ product }) => {
-  const { addCartItemMutation, removeCartItemMutation, cartQuery } = useCart();
-  const { showNotification } = usePersonalizationNotification();
-  const [wasAdded, setWasAdded] = useState(false);
-
-  const cartItem: UserCartItem | undefined = cartQuery.data?.data.find((item) => item.productId === product.id);
-
-  const addToCart = async () => {
-    await addCartItemMutation.mutateAsync({ productId: product.id });
-    if (!cartItem) {
-      showNotification('Product added to cart!');
-    }
-    setWasAdded(true);
-  };
-
-  const removeFromCart = async () => {
-    await removeCartItemMutation.mutateAsync({ itemId: cartItem?.id });
-  };
-
-  useDebounce(
-    () => {
-      if (wasAdded) {
-        setWasAdded(false);
-      }
-    },
-    1000,
-    [wasAdded],
-  );
-
-  return (
-    <div className={styles.productCard}>
-      <Image
-        src={product.image}
-        alt={product.name}
-        sizes="100vw"
-        width={250}
-        height={172}
-        className={styles.productCardImage}
-      />
-      <div className={styles.productCardBody}>
-        <div>
-          <div className={styles.productCardName}>{product.name}</div>
-          <div className={styles.productCardSize}>Big</div>
-        </div>
-        <div className={styles.productCardBodyBottomRow}>
-          <div className={styles.productCardPrice}>${product.price.toFixed(2)}</div>
-          <div className={styles.productCardButtons}>
-            {cartItem ? (
-              <ItemCounter count={cartItem.count} addItem={addToCart} removeItem={removeFromCart} />
-            ) : (
-              <Button size="small" onClick={addToCart}>
-                {addCartItemMutation.isLoading ? 'Adding...' : 'Add to cart'}
-              </Button>
-            )}
-
-            <Button size="small" outlined disabled className={styles.addToFavorites}>
-              <Image src={HeartIcon} alt="Add to favorites" />
-            </Button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
+import { Search, SearchHistory } from '../../client/components/personalization/searchComponents';
+import { ProductCard } from '../../client/components/personalization/productCard';
 
 export default function Index({ embed }: CustomPageProps) {
   const { enqueueSnackbar } = useSnackbar();
 
   const { isLoading: isFpDataLoading, data } = useVisitorData({ extendedResult: true });
 
-  // const [didAcknowledge, setDidAcknowledge] = useSessionStorage('didAcknowledgePersonalizationUseCaseWarning', false);
+  const [didAcknowledge, setDidAcknowledge] = useSessionStorage('didAcknowledgePersonalizationUseCaseWarning', false);
   const [search, setSearch] = useState('');
   const [searchQuery, setSearchQuery] = useState('');
   const [userWelcomed, setUserWelcomed] = useState(false);
@@ -230,7 +87,7 @@ export default function Index({ embed }: CustomPageProps) {
 
   return (
     <>
-      {/* <Dialog open={!didAcknowledge}>
+      <Dialog open={!didAcknowledge}>
         <DialogTitle>Heads up!</DialogTitle>
         <DialogContent>
           <DialogContentText whiteSpace="pre-line">
@@ -245,7 +102,7 @@ export default function Index({ embed }: CustomPageProps) {
         <DialogActions>
           <Button onClick={() => setDidAcknowledge(true)}>Okay, I understand</Button>
         </DialogActions>
-      </Dialog> */}
+      </Dialog>
       <UseCaseWrapper useCase={USE_CASES.personalization} embed={embed} contentSx={{ maxWidth: 'none' }}>
         <div className={styles.twoColumnContainer}>
           <div className={styles.leftColumn}>
